backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.10.4
SQLAlchemy==2.0.36
httpx==0.27.2
